ExecuteNonQuery with params.
