+++ src/Query/Providers/IQueryProvider.cs
@@ -165,6 +165,17 @@ namespace Kros.KORM.Query
         Task<int> ExecuteNonQueryAsync(string query);
 
         /// <summary>
+        /// Asynchronously executes arbitrary query.
+        /// </summary>
+        /// <param name="query">Arbitrary SQL query. It should not be SELECT query.</param>
+        /// <param name="parameters">List of query parameters.</param>
+        /// <returns>
+        /// A task that represents the asynchronous operation. The task result contains the
+        /// numbers of affected rows.
+        /// </returns>
+        Task<int> ExecuteNonQueryAsync(string query, params object[] parameters);
+
+        /// <summary>
         /// Asynchronously executes arbitrary query with parameters.
         /// </summary>
         /// <param name="query">Arbitrary SQL query. It should not be SELECT query.</param>
