+++ tests/Kros.KORM.UnitTests/Query/Sql/LinqTranslatorTestBase.cs
@@ -246,6 +246,11 @@ namespace Kros.KORM.UnitTests.Query.Sql
                 throw new NotImplementedException();
             }
 
+            public Task<int> ExecuteNonQueryAsync(string query, params object[] parameters)
+            {
+                throw new NotImplementedException();
+            }
+
             public Task<int> ExecuteNonQueryAsync(string query, CommandParameterCollection parameters)
             {
                 throw new NotImplementedException();
